requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
SQLAlchemy>=2.0
psycopg2-binary>=2.9
python-dotenv>=1.0
PyJWT>=2.8
requests>=2.31
kivy>=2.3
sentence-transformers>=2.6
faiss-cpu>=1.7
scikit-learn>=1.3
ollama>=0.1
